requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
debugpy==1.5.1
decorator==5.1.0
Django==4.0
django-lockdown==4.0.0
entrypoints==0.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==5.1.1
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
python-decouple==3.5
Werkzeug==2.0.2
django-extensions==3.1.5
python-decouple==3.5
